Update prisma schema and improve endpoints
- Updated the prisma schema to cascade on delete for blocks table when
  a user is deleted

- Made improvements to several endpoint routes and controllers to fetch
  more specific data that the client requires

+++ server/src/controllers/friendships-controller.js
@@ -1,6 +1,12 @@
 import prisma from '../../db/pool.js';
 import newError from '../util/newError.js';
 
+const friendshipExists = async (id) => {
+    return await prisma.friendship.findUnique({
+        where: { id },
+    });
+}
+
 export default {
     // ----- GET -----
     getFriendships: async (userId) => {
@@ -60,30 +66,62 @@ export default {
         });
         if(block) throw newError("Cannot add friend while blocked", 403);
 
-        return await prisma.friendship.create({
-            data: { requesterId, receiverId, }
+        // Check that a friendship between the two users does not already exist
+        const friendshipExists = await prisma.friendship.findFirst({
+            where: {
+                OR: [
+                    { requesterId, receiverId },
+                    { requesterId: receiverId, receiverId: requesterId }
+                ]
+            }
         });
+        if(friendshipExists) throw newError("This user has already sent you a request!", 400);
+
+        // Create the friendship
+        const created = await prisma.friendship.create({
+            data: { requesterId, receiverId },
+            include: {
+                requester: { select: { id: true, username: true, lastSeen: true } },
+                receiver: { select: { id: true, username: true, lastSeen: true } },
+            },
+        });
+
+        // Return normalized friendship object
+        const friend = created.requesterId === requesterId ? created.receiver : created.requester;
+        return {
+            id: created.id,
+            friend,
+            accepted: created.accepted,
+            createdAt: created.createdAt,
+            isRequester: true
+        };
     },
 
     // ----- PUT -----
     acceptFriendship: async (id, userId) => {
-        try {
-            return await prisma.friendship.update({
-                where: { id, receiverId: userId },
-                data: { accepted: true }
-            });
-        }
-        // Prevent requesters from accepting the request themselves
-        catch (err) {
-            if(err.code === 'P2025') {
-                throw newError("You are not authorized to accept this request", 403);
-            }
-            throw err;
-        }
+        // Check that the friend request exists
+        const request = await friendshipExists(id);
+        if (!request) throw newError("Friend request not found", 404);
+
+        // Ensure the current user is the receiver
+        if (request.receiverId !== userId)
+            throw newError("You are not authorized to accept this request", 403);
+
+        // Accept the request
+        return await prisma.friendship.update({
+            where: { id },
+            data: { accepted: true },
+        });
     },
+
 
     // ----- DELETE -----
     deleteFriendship: async (id) => {
+        // Check that the friend request exists
+        const request = await friendshipExists(id);
+        if (!request) throw newError("Friend request not found", 404);
+
+        // Delete friendship
         return await prisma.friendship.delete({
             where: { id }
         });
